+++ lib/rubocop/cop/correctors/empty_line_corrector.rb
@@ -9,9 +9,9 @@ module RuboCop
           offense_style, range = node
           lambda do |corrector|
             case offense_style
-            when :no_empty_lines then
+            when :no_empty_lines
               corrector.remove(range)
-            when :empty_lines then
+            when :empty_lines
               corrector.insert_before(range, "\n")
             end
           end